+++ lib/widgets/app_bar/journal_sliver_appbar.dart
@@ -41,7 +41,6 @@ class JournalSliverAppBar extends StatelessWidget {
                     ),
                     text: snapshot.match,
                     onChanged: cubit.setSearchString,
-                    hintText: 'Search...',
                   ),
                   Padding(
                     padding: const EdgeInsets.symmetric(horizontal: 5),