+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 - Category selection for dashboards
 
+### Changed:
+- Improved search field
+- Improved dashboards page header
+
 ## [0.8.314] - 2023-04-11
 ### Changed:
 - Material icons in bottom nav